+++ examples/api-authentication/client/scripts/auth0-angular.js
@@ -62,6 +62,7 @@
     this.$q = $q;
     this.urlBase64Decode = urlBase64Decode;
     this.delegatedTokens = {};
+    this.profile = {};
   }
   Auth0Wrapper.prototype = {};
   Auth0Wrapper.prototype.parseHash = function (locationHash, callback) {
@@ -169,8 +170,8 @@
       }
       that._serialize(id_token, access_token, state);
       that._deserialize();
-      that.getProfile(id_token).then(function (profile) {
-        that.$rootScope.$broadcast(AUTH_EVENTS.loginSuccess, profile);
+      that.getProfile(id_token).then(function () {
+        that.$rootScope.$broadcast(AUTH_EVENTS.loginSuccess, that.profile);
       });
     };
     var that = this;
@@ -202,8 +203,15 @@
         if (err) {
           return deferred.reject(err);
         }
-        that.profile = profile;
-        deferred.resolve(profile);
+        // Cleanup old keys
+        Object.keys(that.profile).forEach(function (key) {
+          delete that.profile[key];
+        });
+        // Add new keys
+        Object.keys(profile).forEach(function (key) {
+          that.profile[key] = profile[key];
+        });
+        deferred.resolve(that.profile);
       });
     this.auth0Lib.getProfile(token, wrappedCallback);
     return deferred.promise;
@@ -218,18 +226,30 @@
     '$provide',
     function ($provide) {
       var auth0Wrapper;
-      this.init = function (options) {
+      /**
+     * Initializes the auth service.
+     *
+     * this.init(options, [Auth0Constructor])
+     *
+     * @param options           object   Options for auth0.js or widget
+     * @param Auth0Constructor  function (optional) constructor to create auth0Lib
+     */
+      this.init = function (options, Auth0Constructor) {
         var auth0Lib;
         if (options.callbackOnLocationHash === undefined) {
           options.callbackOnLocationHash = true;
         }
-        // User has included widget
-        if (typeof Auth0Widget !== 'undefined') {
-          auth0Lib = new Auth0Widget(options);
-        } else if (typeof Auth0 !== 'undefined') {
-          auth0Lib = new Auth0(options);
+        if (Auth0Constructor) {
+          auth0Lib = new Auth0Constructor(options);
         } else {
-          throw new Error('You need to add Auth0Widget or Auth0.js dependency');
+          // User has included Auth0 widget
+          if (typeof Auth0Widget !== 'undefined') {
+            auth0Lib = new Auth0Widget(options);
+          } else if (typeof Auth0 !== 'undefined') {
+            auth0Lib = new Auth0(options);
+          } else {
+            throw new Error('You need to add Auth0Widget or Auth0.js dependency');
+          }
         }
         $provide.value('auth0Lib', auth0Lib);
       };
@@ -263,11 +283,11 @@
         var result = auth.parseHash($window.location.hash);
         if (result && result.id_token) {
           // this is only used when using redirect mode
-          auth.getProfile(result.id_token).then(function (profile) {
+          auth.getProfile(result.id_token).then(function () {
             auth._serialize(result.id_token, result.access_token, result.state);
             // this will rehydrate the "auth" object with the profile stored in $cookieStore
             auth._deserialize();
-            $rootScope.$broadcast(AUTH_EVENTS.loginSuccess, profile);
+            $rootScope.$broadcast(AUTH_EVENTS.loginSuccess, auth.profile);
           }, function (err) {
             // this will rehydrate the "auth" object with the profile stored in $cookieStore
             auth._deserialize();